refactor: split admin.ts into admin/ directory (system, users, parseLogs, apiKeys, logging, parserSettings)
