+++ backend/src/routes/admin/parserSettings.ts
@@ -1,0 +1,262 @@
+import { Router, Response, NextFunction } from 'express';
+import { z } from 'zod';
+import { AuthRequest, requireAdmin } from '../../middleware/auth';
+import { prisma } from '../../db';
+import { decryptApiKey, encryptApiKey } from '../../utils/encryption';
+import logger from '../../utils/logger';
+
+interface ParserSettingsUpdateData {
+  globalOpenaiApiKey?: string | null;
+  globalClaudeApiKey?: string | null;
+  allowUserApiKeys?: boolean;
+  requireUserApiKeys?: boolean;
+  defaultVisionParser?: string;
+  defaultTextParser?: string;
+}
+
+const parserSettingsSchema = z.object({
+  globalOpenaiApiKey: z.string().nullable().optional(),
+  globalClaudeApiKey: z.string().nullable().optional(),
+  allowUserApiKeys: z.boolean().optional(),
+  requireUserApiKeys: z.boolean().optional(),
+  defaultVisionParser: z.string().optional(),
+  defaultTextParser: z.string().optional(),
+});
+
+// Training configuration schema
+const trainingConfigSchema = z.object({
+  trainingModelOutputDir: z.string().optional().nullable().refine(
+    (val) => !val || (val.length > 0 && val.length <= 500 && /^[a-zA-Z0-9/._-]+$/.test(val)),
+    { message: 'Invalid path format (max 500 chars, alphanumeric, /, ., _, - only)' }
+  ),
+  trainingEmailModelName: z.string().optional().nullable().refine(
+    (val) => !val || (val.length > 0 && val.length <= 100 && /^[a-zA-Z0-9._-]+$/.test(val)),
+    { message: 'Invalid model name format (max 100 chars, alphanumeric, ., _, - only)' }
+  ),
+  trainingVisionModelName: z.string().optional().nullable().refine(
+    (val) => !val || (val.length > 0 && val.length <= 100 && /^[a-zA-Z0-9._-]+$/.test(val)),
+    { message: 'Invalid model name format (max 100 chars, alphanumeric, ., _, - only)' }
+  ),
+});
+
+const router = Router();
+
+// Get admin parser settings
+router.get('/parser-settings', async (req: AuthRequest, res: Response, next: NextFunction) => {
+  try {
+    // Get or create admin settings (ID is always 1 for singleton)
+    let adminSettings = await prisma.adminSettings.findFirst();
+
+    if (!adminSettings) {
+      // Create default admin settings if they don't exist
+      adminSettings = await prisma.adminSettings.create({
+        data: {
+          allowUserApiKeys: true,
+          requireUserApiKeys: false,
+          defaultVisionParser: 'auto',
+          defaultTextParser: 'auto',
+        },
+      });
+    }
+
+    // Return settings (API keys are decrypted for frontend)
+    res.json({
+      globalOpenaiApiKey: decryptApiKey(adminSettings.globalOpenaiApiKey) || undefined,
+      globalClaudeApiKey: decryptApiKey(adminSettings.globalClaudeApiKey) || undefined,
+      globalAirlabsApiKey: decryptApiKey(adminSettings.globalAirlabsApiKey) || undefined,
+      globalAviationstackApiKey: decryptApiKey(adminSettings.globalAviationstackApiKey) || undefined,
+      globalOpenskyClientId: decryptApiKey(adminSettings.globalOpenskyClientId) || undefined,
+      globalOpenskyClientSecret: decryptApiKey(adminSettings.globalOpenskyClientSecret) || undefined,
+      globalOpenskyUsername: decryptApiKey(adminSettings.globalOpenskyUsername) || undefined,
+      globalOpenskyPassword: decryptApiKey(adminSettings.globalOpenskyPassword) || undefined,
+      allowUserApiKeys: adminSettings.allowUserApiKeys,
+      requireUserApiKeys: adminSettings.requireUserApiKeys,
+      allowUserFlightApiKeys: adminSettings.allowUserFlightApiKeys,
+      requireUserFlightApiKeys: adminSettings.requireUserFlightApiKeys,
+      defaultVisionParser: adminSettings.defaultVisionParser,
+      defaultTextParser: adminSettings.defaultTextParser,
+    });
+  } catch (error) {
+    next(error);
+  }
+});
+
+// Update admin parser settings
+router.put('/parser-settings', async (req: AuthRequest, res: Response, next: NextFunction) => {
+  try {
+    const {
+      globalOpenaiApiKey,
+      globalClaudeApiKey,
+      allowUserApiKeys,
+      requireUserApiKeys,
+      defaultVisionParser,
+      defaultTextParser,
+    } = parserSettingsSchema.parse(req.body);
+
+    // Get or create admin settings
+    let adminSettings = await prisma.adminSettings.findFirst();
+
+    const updateData: ParserSettingsUpdateData = {};
+
+    // Only update fields that are provided
+    // Encrypt API keys before storing
+    if (globalOpenaiApiKey !== undefined) {
+      updateData.globalOpenaiApiKey = encryptApiKey(globalOpenaiApiKey);
+    }
+    if (globalClaudeApiKey !== undefined) {
+      updateData.globalClaudeApiKey = encryptApiKey(globalClaudeApiKey);
+    }
+    if (allowUserApiKeys !== undefined) {
+      updateData.allowUserApiKeys = allowUserApiKeys;
+    }
+    if (requireUserApiKeys !== undefined) {
+      updateData.requireUserApiKeys = requireUserApiKeys;
+    }
+    if (defaultVisionParser !== undefined) {
+      updateData.defaultVisionParser = defaultVisionParser;
+    }
+    if (defaultTextParser !== undefined) {
+      updateData.defaultTextParser = defaultTextParser;
+    }
+
+    if (adminSettings) {
+      // Update existing settings
+      adminSettings = await prisma.adminSettings.update({
+        where: { id: adminSettings.id },
+        data: updateData,
+      });
+    } else {
+      // Create new settings with provided data
+      adminSettings = await prisma.adminSettings.create({
+        data: {
+          globalOpenaiApiKey: encryptApiKey(globalOpenaiApiKey),
+          globalClaudeApiKey: encryptApiKey(globalClaudeApiKey),
+          allowUserApiKeys: allowUserApiKeys ?? true,
+          requireUserApiKeys: requireUserApiKeys ?? false,
+          allowUserFlightApiKeys: true,
+          requireUserFlightApiKeys: false,
+          defaultVisionParser: defaultVisionParser || 'auto',
+          defaultTextParser: defaultTextParser || 'auto',
+        },
+      });
+    }
+
+    res.json({
+      message: 'Parser settings updated successfully',
+      settings: {
+        globalOpenaiApiKey: decryptApiKey(adminSettings.globalOpenaiApiKey) || undefined,
+        globalClaudeApiKey: decryptApiKey(adminSettings.globalClaudeApiKey) || undefined,
+        allowUserApiKeys: adminSettings.allowUserApiKeys,
+        requireUserApiKeys: adminSettings.requireUserApiKeys,
+        defaultVisionParser: adminSettings.defaultVisionParser,
+        defaultTextParser: adminSettings.defaultTextParser,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+});
+
+// Get training configuration
+router.get('/training-config', requireAdmin, async (req: AuthRequest, res: Response, next: NextFunction) => {
+  try {
+    const adminSettings = await prisma.adminSettings.findFirst();
+    const trainingModelOutputDir = adminSettings?.trainingModelOutputDir
+      || process.env.TRAINING_MODEL_OUTPUT_DIR || './data/training/models';
+    const trainingEmailModelName = adminSettings?.trainingEmailModelName
+      || process.env.TRAINING_EMAIL_MODEL_NAME || 'travstats-email-custom';
+    const trainingVisionModelName = adminSettings?.trainingVisionModelName
+      || process.env.TRAINING_VISION_MODEL_NAME || 'travstats-vision-custom';
+
+    res.json({
+      trainingModelOutputDir: adminSettings?.trainingModelOutputDir || null,
+      trainingEmailModelName: adminSettings?.trainingEmailModelName || null,
+      trainingVisionModelName: adminSettings?.trainingVisionModelName || null,
+      // Current effective values (from ENV if not set in admin)
+      currentTrainingModelOutputDir: trainingModelOutputDir,
+      currentTrainingEmailModelName: trainingEmailModelName,
+      currentTrainingVisionModelName: trainingVisionModelName,
+      // ENV fallback values
+      envTrainingModelOutputDir: process.env.TRAINING_MODEL_OUTPUT_DIR || './data/training/models',
+      envTrainingEmailModelName: process.env.TRAINING_EMAIL_MODEL_NAME || 'travstats-email-custom',
+      envTrainingVisionModelName: process.env.TRAINING_VISION_MODEL_NAME || 'travstats-vision-custom',
+    });
+  } catch (error) {
+    next(error);
+  }
+});
+
+// Update training configuration
+router.put('/training-config', requireAdmin, async (req: AuthRequest, res: Response, next: NextFunction) => {
+  try {
+    const payload = trainingConfigSchema.parse(req.body);
+
+    const updateData: {
+      trainingModelOutputDir?: string | null;
+      trainingEmailModelName?: string | null;
+      trainingVisionModelName?: string | null;
+    } = {};
+
+    if (payload.trainingModelOutputDir !== undefined) {
+      updateData.trainingModelOutputDir = payload.trainingModelOutputDir || null;
+    }
+    if (payload.trainingEmailModelName !== undefined) {
+      updateData.trainingEmailModelName = payload.trainingEmailModelName || null;
+    }
+    if (payload.trainingVisionModelName !== undefined) {
+      updateData.trainingVisionModelName = payload.trainingVisionModelName || null;
+    }
+
+    let adminSettings = await prisma.adminSettings.findFirst();
+
+    if (adminSettings) {
+      adminSettings = await prisma.adminSettings.update({
+        where: { id: adminSettings.id },
+        data: updateData,
+      });
+    } else {
+      adminSettings = await prisma.adminSettings.create({
+        data: {
+          allowUserApiKeys: true,
+          requireUserApiKeys: false,
+          defaultVisionParser: 'auto',
+          defaultTextParser: 'auto',
+          ...updateData,
+        },
+      });
+    }
+
+    logger.info({
+      operation: 'training_config_updated',
+      message: 'Training configuration updated',
+      context: {
+        userId: req.userId,
+        settings: updateData,
+      },
+    });
+
+    // Get updated effective values
+    const currentTrainingModelOutputDir = adminSettings.trainingModelOutputDir
+      || process.env.TRAINING_MODEL_OUTPUT_DIR || './data/training/models';
+    const currentTrainingEmailModelName = adminSettings.trainingEmailModelName
+      || process.env.TRAINING_EMAIL_MODEL_NAME || 'travstats-email-custom';
+    const currentTrainingVisionModelName = adminSettings.trainingVisionModelName
+      || process.env.TRAINING_VISION_MODEL_NAME || 'travstats-vision-custom';
+
+    res.json({
+      message: 'Training configuration updated successfully',
+      settings: {
+        trainingModelOutputDir: adminSettings.trainingModelOutputDir,
+        trainingEmailModelName: adminSettings.trainingEmailModelName,
+        trainingVisionModelName: adminSettings.trainingVisionModelName,
+        currentTrainingModelOutputDir,
+        currentTrainingEmailModelName,
+        currentTrainingVisionModelName,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+});
+
+export default router;
